+++ src/PackedImageData.cpp
@@ -14,7 +14,7 @@ using namespace canvas;
 
 bool PackedImageData::etc1_initialized = false;
 
-PackedImageData::PackedImageData(InternalFormat _format, unsigned int _levels, const ImageData & input)
+PackedImageData::PackedImageData(InternalFormat _format, unsigned short _levels, const ImageData & input)
   : format(_format), width(input.getWidth()), height(input.getHeight()), levels(_levels)
 {
   if (format == NO_FORMAT) {
@@ -27,7 +27,7 @@ PackedImageData::PackedImageData(InternalFormat _format, unsigned int _levels, c
     }
   }
 
-  unsigned int num_channels = input.getNumChannels();
+  unsigned short num_channels = input.getNumChannels();
 
   size_t s = calculateSize();
   data = std::unique_ptr<unsigned char[]>(new unsigned char[s]);
@@ -41,7 +41,7 @@ PackedImageData::PackedImageData(InternalFormat _format, unsigned int _levels, c
     unsigned int offset = fs.apply(input, data.get());
     if (levels >= 2) {
       auto img = input.scale((input.getWidth() + 1) / 2, (input.getHeight() + 1) / 2);
-      for (int l = 1; l < levels; l++) {
+      for (unsigned int l = 1; l < levels; l++) {
 	offset += fs.apply(*img, data.get() + offset);
 	if (l + 1 < levels) {
 	  img = img->scale((img->getWidth() + 1) / 2, (img->getHeight() + 1) / 2);
@@ -75,7 +75,7 @@ PackedImageData::PackedImageData(InternalFormat _format, unsigned int _levels, c
 	unsigned char g = num_channels >= 2 ? input_data[input_offset++] : r;
 	unsigned char b = num_channels >= 3 ? input_data[input_offset++] : g;
 	unsigned char a = (num_channels >= 4 ? input_data[input_offset++] : 0xff) >> 4;
-	int lum = ((r + g + b) / 3) >> 4;
+	unsigned int lum = ((r + g + b) / 3) >> 4;
 	if (lum >= 16) lum = 15;
 	*output_data++ = (a << 4) | lum;
       }
@@ -86,7 +86,7 @@ PackedImageData::PackedImageData(InternalFormat _format, unsigned int _levels, c
   }
 }
 
-PackedImageData::PackedImageData(InternalFormat _format, unsigned int _width, unsigned int _height, unsigned int _levels)
+PackedImageData::PackedImageData(InternalFormat _format, unsigned short _width, unsigned short _height, unsigned short _levels)
   : width(_width), height(_height), levels(_levels), format(_format) {
   size_t s = calculateSize();
   data = std::unique_ptr<unsigned char[]>(new unsigned char[s]);
@@ -119,7 +119,7 @@ PackedImageData::PackedImageData(InternalFormat _format, unsigned int _width, un
 
 #if 0
 void
-PackedImageData::createMipmaps(const ImageData & input_data, unsigned int target_levels) const {
+PackedImageData::createMipmaps(const ImageData & input_data, unsigned short target_levels) const {
   auto & fd = getImageFormat(format);
   assert(fd.getBytesPerPixel() == 4);
   assert(!fd.getCompression());
@@ -127,20 +127,20 @@ PackedImageData::createMipmaps(const ImageData & input_data, unsigned int target
   size_t target_size = calculateOffset(target_levels);
   std::unique_ptr<unsigned char[]> output_data(new unsigned char[target_size]);
   memcpy(output_data.get(), data.get(), calculateOffset(1));
-  unsigned int source_width = width, source_height = height;
-  unsigned int target_width = width / 2, target_height = height / 2;
+  unsigned short source_width = width, source_height = height;
+  unsigned short target_width = width / 2, target_height = height / 2;
   for (unsigned int level = 1; level < target_levels; level++) {
     unsigned char * source_data = output_data.get() + calculateOffset(level - 1);
     unsigned char * target_data = output_data.get() + calculateOffset(level);
-    for (int y = 0; y < target_height; y++) {               
-      for (int x = 0; x < target_width; x++) {
+    for (unsigned int y = 0; y < target_height; y++) {               
+      for (unsigned int x = 0; x < target_width; x++) {
 	unsigned int source_offset = (2 * y * source_width + 2 * x) * 4;
 	unsigned int target_offset = (y * target_width + x) * 4;
 	for (unsigned int c = 0; c < 4; c++) {
-	  target_data[target_offset] = ((int)source_data[source_offset]  + 
-					(int)source_data[source_offset + 4]  + 
-					(int)source_data[source_offset + 4 + 4 * source_width] +
-					(int)source_data[source_offset + 4 * source_width]) / 4; 
+	  target_data[target_offset] = ((unsigned int)source_data[source_offset]  + 
+					(unsigned int)source_data[source_offset + 4]  + 
+					(unsigned int)source_data[source_offset + 4 + 4 * source_width] +
+					(unsigned int)source_data[source_offset + 4 * source_width]) / 4; 
 	  source_offset++;
 	  target_offset++;
 	}
@@ -170,41 +170,41 @@ ImageData::convert(InternalFormat target_format) const {
     }
     assert((width & 3) == 0);
     assert((height & 3) == 0);
-    unsigned int target_width = width, target_height = height;
+    unsigned short target_width = width, target_height = height;
     unsigned int target_size = calculateSize(target_width, target_height, levels, target_format);
     std::unique_ptr<unsigned char[]> output_data(new unsigned char[target_size]);
     unsigned char input_block[4*4*8];
     unsigned int target_offset = 0;
     for (unsigned int level = 0; level < levels; level++) {
       unsigned int rows = (target_height + 3) / 4, cols = (target_width + 3) / 4;
-      int base_source_offset = calculateOffset(level);
+      unsigned int base_source_offset = calculateOffset(level);
       // cerr << "compressing texture, level " << level << ", rows = " << rows << ", cols = " << cols << "\n";
       for (unsigned int row = 0; row < rows; row++) {
 	for (unsigned int col = 0; col < cols; col++) {
 	  for (unsigned int y = 0; y < 4; y++) {
 	    for (unsigned int x = 0; x < 4; x++) {
-	      int source_offset = base_source_offset + ((row * 4 + y) * target_width + col * 4 + x) * fd.getBytesPerPixel();
+	      unsigned int source_offset = base_source_offset + ((row * 4 + y) * target_width + col * 4 + x) * fd.getBytesPerPixel();
 	      unsigned char r = data[source_offset++];
 	      unsigned char g = fd.getBytesPerPixel() >= 2 ? data[source_offset++] : r;
 	      unsigned char b = fd.getBytesPerPixel() >= 3 ? data[source_offset++] : g;
 	      unsigned char a = fd.getBytesPerPixel() >= 4 ? data[source_offset++] : 0xff;
 	      if (target_fd.getCompression() == ImageFormat::ETC1) {
-		int offset = (y * 4 + x) * 4;
+		unsigned int offset = (y * 4 + x) * 4;
 		input_block[offset++] = r;
 		input_block[offset++] = g;
 		input_block[offset++] = b;
 		input_block[offset++] = 255; // data[source_offset++];
 	      } else if (target_fd.getCompression() == ImageFormat::DXT1) {
-		int offset = (y * 4 + x) * 4;
+		unsigned int offset = (y * 4 + x) * 4;
 		input_block[offset++] = b;
 		input_block[offset++] = g;
 		input_block[offset++] = r;
 		input_block[offset++] = 255; // data[source_offset++];
 	      } else if (target_fd.getCompression() == ImageFormat::RGTC1) {
-		int offset = y * 4 + x;
+		unsigned int offset = y * 4 + x;
 		input_block[offset] = r;
 	      } else {
-		int offset = y * 4 + x;
+		unsigned int offset = y * 4 + x;
 		input_block[offset] = r;
 		input_block[offset + 16] = a;
 	      }
@@ -258,7 +258,7 @@ ImageData::convert(InternalFormat target_format) const {
 	unsigned char g = fd.getBytesPerPixel() >= 1 ? data[input_offset++] : r;
 	unsigned char b = fd.getBytesPerPixel() >= 2 ? data[input_offset++] : g;
 	unsigned char a = fd.getBytesPerPixel() >= 3 ? data[input_offset++] : 0xff;
-	int lum = (r + g + b) / 3;
+	unsigned int lum = (r + g + b) / 3;
 	if (lum >= 255) lum = 255;
 	*output_data++ = (a << 8) | lum;
       }
@@ -282,36 +282,36 @@ ImageData::convert(InternalFormat target_format) const {
 }
 
 std::unique_ptr<ImageData>
-ImageData::scale(unsigned int target_base_width, unsigned int target_base_height, unsigned int target_levels) const {
+ImageData::scale(unsigned short target_base_width, unsigned short target_base_height, unsigned short target_levels) const {
   auto & fd = getImageFormat(format);
   assert(!fd.getCompression());
   size_t target_size = calculateOffset(target_base_width, target_base_height, target_levels, format);
   // cerr << "scaling to " << target_base_width << " " << target_base_height << " " << target_levels << " => " << target_size << " bytes\n";
   std::unique_ptr<unsigned char[]> output_data(new unsigned char[target_size]);
-  unsigned int target_width = target_base_width, target_height = target_base_height;
+  unsigned short target_width = target_base_width, target_height = target_base_height;
   stbir_resize_uint8(data.get(), getWidth(), getHeight(), 0, output_data.get(), target_width, target_height, 0, fd.getBytesPerPixel());
 
   if (target_levels > 1) {
     cerr << "creating mipmaps for format, channels = " << fd.getNumChannels() << ", bpp = " << fd.getBytesPerPixel() << endl;
 
-    unsigned int source_width = target_width, source_height = target_height;
+    unsigned short source_width = target_width, source_height = target_height;
     target_width /= 2;
     target_height /= 2;
 
-    unsigned int nch = fd.getBytesPerPixel();
+    unsigned short nch = fd.getBytesPerPixel();
     
     for (unsigned int level = 1; level < target_levels; level++) {
       unsigned char * source_data = output_data.get() + calculateOffset(target_base_width, target_base_height, level - 1, format);
       unsigned char * target_data = output_data.get() + calculateOffset(target_base_width, target_base_height, level, format);
-      for (int y = 0; y < target_height; y++) {               
-	for (int x = 0; x < target_width; x++) {
+      for (unsigned int y = 0; y < target_height; y++) {               
+	for (unsigned int x = 0; x < target_width; x++) {
 	  unsigned int source_offset = (2 * y * source_width + 2 * x) * nch;
 	  unsigned int target_offset = (y * target_width + x) * nch;
 	  for (unsigned int c = 0; c < nch; c++) {
-	    target_data[target_offset] = ((int)source_data[source_offset]  + 
-					  (int)source_data[source_offset + nch]  + 
-					  (int)source_data[source_offset + nch + nch * source_width] +
-					  (int)source_data[source_offset + nch * source_width]) / 4; 
+	    target_data[target_offset] = ((unsigned int)source_data[source_offset]  + 
+					  (unsigned int)source_data[source_offset + nch]  + 
+					  (unsigned int)source_data[source_offset + nch + nch * source_width] +
+					  (unsigned int)source_data[source_offset + nch * source_width]) / 4; 
 	    source_offset++;
 	    target_offset++;
 	  }
